requirements.txt
asgiref==3.7.2
Django==4.2.11
django-ckeditor==6.7.1
django-jalali==6.0.1
django-js-asset==2.2.0
jdatetime==4.1.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
